+++ packages/arch3-core/src/archwayclient.ts
@@ -2,11 +2,11 @@ import { Coin } from '@cosmjs/amino';
 import { CosmWasmClient } from '@cosmjs/cosmwasm-stargate';
 import {
   Tendermint37Client,
-  TendermintClient,
   HttpEndpoint,
   RpcClient,
   HttpBatchClient,
-  HttpBatchClientOptions
+  HttpBatchClientOptions,
+  CometClient
 } from '@cosmjs/tendermint-rpc';
 
 import { IArchwayQueryClient, createArchwayQueryClient } from './queryclient';
@@ -26,18 +26,18 @@ import {
 export class ArchwayClient extends CosmWasmClient implements IArchwayQueryClient {
   private readonly archwayQueryClient: IArchwayQueryClient;
 
-  protected constructor(tmClient: TendermintClient | undefined) {
-    super(tmClient);
-    this.archwayQueryClient = createArchwayQueryClient(tmClient);
+  protected constructor(cometClient: CometClient | undefined) {
+    super(cometClient);
+    this.archwayQueryClient = createArchwayQueryClient(cometClient);
   }
 
   /**
-   * Creates an instance by connecting to the given Tendermint RPC endpoint.
+   * Creates an instance by connecting to the given Tendermint/Comet RPC endpoint.
    *
    * @param endpoint - String URL of the RPC endpoint to connect or an {@link HttpEndpoint} object.
    * @returns An {@link ArchwayClient} connected to the endpoint.
    *
-   * @see Use {@link ArchwayClient.create} if you need Tendermint 0.37 support.
+   * @see Use {@link ArchwayClient.create} if you need Comet 0.38 support.
    */
   public static override async connect(endpoint: string | HttpEndpoint): Promise<ArchwayClient> {
     const tmClient = await Tendermint37Client.connect(endpoint);
@@ -45,7 +45,7 @@ export class ArchwayClient extends CosmWasmClient implements IArchwayQueryClient
   }
 
   /**
-   * Creates an instance by connecting to the given Tendermint RPC endpoint using an {@link HttpBatchClient} to batch
+   * Creates an instance by connecting to the given Tendermint/Comet RPC endpoint using an {@link HttpBatchClient} to batch
    * multiple requests and reduce queries to the server.
    *
    * @param endpoint - String URL of the RPC endpoint to connect or an {@link HttpEndpoint} object.
@@ -61,14 +61,14 @@ export class ArchwayClient extends CosmWasmClient implements IArchwayQueryClient
   }
 
   /**
-   * Creates an instance from a manually created Tendermint client.
+   * Creates an instance from a manually created Comet client.
    *
-   * @param tmClient - A Tendermint client for a given endpoint.
+   * @param cometClient - A Comet client for a given endpoint.
    * @returns An {@link ArchwayClient} connected to the endpoint.
    */
   /* eslint-disable-next-line @typescript-eslint/require-await */
-  public static override async create(tmClient: TendermintClient): Promise<ArchwayClient> {
-    return new ArchwayClient(tmClient);
+  public static override async create(cometClient: CometClient): Promise<ArchwayClient> {
+    return new ArchwayClient(cometClient);
   }
 
   public async getBlockRewardsTracking(): Promise<BlockTracking> {