+++ packages/arch3-proto/scripts/codegen.js
@@ -50,8 +50,10 @@ telescope({
         useDeepPartial: false,
         useExact: false,
         timestamp: 'timestamp',
-        duration: 'duration'
+        duration: 'duration',
+        useTelescopeGeneratedType: true
       },
+      enableRegistryLoader: false,
       methods: {
         toJSON: true,
         fromJSON: true